requirements.txt
streamlit
numpy
opencv-python
Pillow
scikit-image
scikit-learn
joblib